Add in one legitimate type-narrowing issue that MSVC picked up :)

+++ tests/MemoryCard.cpp
@@ -17,7 +17,7 @@ static std::array<Byte, SIZE> generate_random_bytes() {
     // N.B: Can't use uint8_t for type as generator doesn't support char types
     std::uniform_int_distribution<std::uint16_t> prng(0x0000, 0x00FF);
     for (auto& d : data) {
-        d = prng(engine);
+        d = (Byte)prng(engine);
     }
     return data;
 }
